pyproject.toml
[project]
name = "tjm-vision-project"
version = "1.0.0"
requires-python = ">=3.13"
dependencies = [
    "pyautogui",
    "requests",
    "PyGetWindow",
    "botcity-framework-core",
    "pyperclip",
]
